# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Job processing backend API for Atlan Collect",
        "title": "Job submitting backend",
        "contact": {},
        "license": {},
        "version": "0.1"
    },
    "paths": {
        "/details/{jobID}": {
            "get": {
                "description": "Job processing backend API for Atlan Collect",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Fetch details about a submitted job",
                "operationId": "details-job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.httpResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.httpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.httpError"
                        }
                    }
                }
            }
        },
        "/halt/{jobID}": {
            "get": {
                "description": "Job processing backend API for Atlan Collect",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Halt a running job",
                "operationId": "halt-job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.httpResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.httpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.httpError"
                        }
                    }
                }
            }
        },
        "/resume/{jobID}": {
            "get": {
                "description": "Job processing backend API for Atlan Collect",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Resume a pause/halted job",
                "operationId": "resume-job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.httpResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.httpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.httpError"
                        }
                    }
                }
            }
        },
        "/stop/{jobID}": {
            "get": {
                "description": "Job processing backend API for Atlan Collect",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Stop a  job",
                "operationId": "stop-job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.httpResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.httpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.httpError"
                        }
                    }
                }
            }
        },
        "/submit": {
            "post": {
                "description": "Job processing backend API for Atlan Collect",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Submit a job for processing",
                "operationId": "submit-job",
                "parameters": [
                    {
                        "description": "Submit a job",
                        "name": "jobRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.JobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.httpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/main.httpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.httpError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.JobRequest": {
            "type": "object",
            "properties": {
                "Type": {
                    "type": "string",
                    "example": "Simple"
                },
                "args": {
                    "type": "object"
                }
            }
        },
        "main.httpError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid JobID"
                },
                "jobID": {
                    "type": "string",
                    "example": "55e75f6c-24f8-49b5-9e62-a268db7370e9"
                }
            }
        },
        "main.httpResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "object"
                },
                "jobID": {
                    "type": "string",
                    "example": "55e75f6c-24f8-49b5-9e62-a268db7370e9"
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        }
    }
}